+++ packages/react/CHANGELOG.md
@@ -9,6 +9,7 @@ description: All notable changes to this project will be documented in this file
 ### Added
 
 - Added `Toggle` component
+- Added `ToggleGroup` component
 
 ### Fixed